🐞 fix(b_drv_rs485中缓冲数据大小变为U32)

+++ bos/hal/inc/b_hal_uart.h
@@ -77,7 +77,7 @@ typedef struct
     bHalUartNumber_t       uart;
     bHalIt_t               it;
     uint8_t               *pbuf;
-    uint16_t               len;
+    uint32_t               len;
     uint16_t               idle_ms;
     bHalUartIdleCallback_t callback;
     void                  *user_data;
